attack-code/requirements.txt
cryptography
pyserial
numpy
matplotlib
